#1166 Phase 1: extract segment_forwarded_tcp_frames_into_prepared to tx/tcp_segmentation.rs
Pure code motion: 281-line function moved from tx/dispatch.rs to a new
sibling tx/tcp_segmentation.rs (use super::*; pub(super) fn ...). Caller
in tx/dispatch.rs imports via use super::tcp_segmentation::....

dispatch.rs shrinks from 1517 to 1236 LOC. No behavior change; cargo
build clean, 974 tests pass, 5/5 flake check on
segment_forwarded_tcp_frames_splits_ipv6_snat_payload_by_mtu.

Plan: docs/pr/1166-tso-extract/plan.md (v4 commit cb6cab25)
Codex round-4: PLAN-READY
